moved endpoints back to instancescontroller
